Separated stats page file from CountEmotionsGraph, so that I could easily re-use the graph and allow access to the data from all of the graphs I plan on creating for the stats page.
